pyproject.toml
[project]
name = "langchain_learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.1.0",
    "langchain-community>=0.3.27",
    "langchain-core>=0.1.23",
    "langchain-google-genai>=1.0.1",
    "langchain-huggingface>=0.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "termcolor>=3.1.0",
    "uvicorn>=0.35.0",
]
